refactor: drop the -actions suffix inside actions/
`actions/import-actions.ts` said it twice. The names are `actions/import`
and `actions/export` now, which the last commit made possible: `import.ts`
was taken until `importBPMN` left `utils/`.
