+++ bpmn-editor/react/src/components/file-import-overlay/file-import-overlay.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef } from 'react';
 import { usePaperScroller, useGraphHistory } from '@joint/react-plus';
-import { importFile } from '../../actions/import-actions';
+import { importFile } from '../../actions/import';
 import './file-import-overlay.css';
 
 /**
